pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
exclude = [
    "test_*.py",
]

[project]
name = "job-notifier"
version = "0.0.1"
requires_python = ">=3.13"
authors = [
    {"name" = "someone", "email" = "someone@example.com"}
]
description = "Job Notifier"
dependencies = [
    "httpx",
    "lxml",
    "markdown",
]
[project.optional-dependencies]
dev = [
    "pre-commit",
    "ipython",
    "pytest",
    "pytest-httpx",
    "coverage",
    "pytest-cov",
    "google-api-python-client",
    "sqlalchemy",
    "psycopg2-binary",
    "python-dotenv",
    "schedule",
]

[tool.coverage.run]
branch = true

[tool.coverage.report]
fail_under = 98
show_missing = true
skip_covered = true

[tool.ruff]
line-length=88

[tool.ruff.lint]
extend-select = ["E501"]
